pyproject.toml
[project]
name = "moiql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.2.1",
    "highway-env>=1.10.1",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "mo-gymnasium>=1.3.1",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "plotly>=6.3.0",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.1",
]
